feat(Flux): add Flux component for traffic monitoring and refactor Info.vue
